ci: add rpcimportable test
