+++ contrib/rpcimportable/rpcimportable_test.go
@@ -1,0 +1,11 @@
+package rpcimportable
+
+import (
+	"testing"
+
+	"github.com/zeta-chain/node/pkg/rpc"
+)
+
+func TestRPCImportable(t *testing.T) {
+	_ = rpc.Clients{}
+}
